# productService/swagger.yml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/product":{"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteProduct","requestBody":{"content":{"application/json":{"schema":{"type":"string","format":"uuid"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/products":{"get":{"tags":["product-controller"],"operationId":"getAllProducts","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/products/{productId}":{"get":{"tags":["product-controller"],"operationId":"updateProductAmount","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"difference","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/product/{productId}":{"get":{"tags":["product-controller"],"operationId":"getProductById","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/product/name/{name}":{"get":{"tags":["product-controller"],"operationId":"findProductsByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/category/{category}":{"get":{"tags":["product-controller"],"operationId":"findProductsByCategory","parameters":[{"name":"category","in":"path","required":true,"schema":{"type":"string","enum":["OCEAN","EARTH","BABY","GIANT"]}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}}},"components":{"schemas":{"Product":{"required":["description","name","price"],"type":"object","properties":{"productId":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"imageLink":{"type":"string"},"category":{"type":"string","enum":["OCEAN","EARTH","BABY","GIANT"]},"material":{"type":"string","enum":["BAUMWOLLE","POLYESTER","WOLLE","FLUFFY"]},"color":{"type":"string","enum":["BLUE","BROWN","ROSE","GREEN","GREY","COLORFUL"]},"size":{"type":"string"},"amount":{"type":"integer","format":"int32"}}}}}}